zOld-Code/step1--Base_FILENAME--a-caselaw/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-renamer-caselaw"
version = "0.2.0"
description = "Batch rename caselaw files (PDF, Word) using Bluebook format with comprehensive metadata extraction and central registry management."
authors = [
  { name = "Dan", email = "dan@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.6"
dependencies = [
  "python-docx>=0.8.11",
  # System-level requirement: pdftotext command (install poppler-utils)
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
caselaw-renamer = "file_renamer.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
